Explicitly close socket after response has entered finished state
In at least two places in the code, the response object is evaluated
on whether it has finished sending data to the client. If the response
has already finished, or enters the finished state after sending data,
the underlying socket should not be left open. Failure to close the
socket in these scenarios cause the remote client to timeout expecting
more data unless it closes the connection on its own.

Fixes: me-no-dev#802